+++ cargo-espflash/src/main.rs
@@ -8,7 +8,6 @@ use cargo_metadata::{Message, MetadataCommand};
 use clap::{Args, CommandFactory, Parser, Subcommand};
 use espflash::{
     Error as EspflashError,
-    bt::check_bootloader,
     cli::{
         self,
         config::Config,
@@ -16,6 +15,7 @@ use espflash::{
         *,
     },
     flasher::FlashSize,
+    image_format::check_idf_bootloader,
     logging::initialize_logger,
     targets::{Chip, XtalFrequency},
     update::check_for_update,
@@ -321,7 +321,7 @@ fn flash(args: FlashArgs, config: &Config) -> Result<()> {
 
     // Check if the ELF contains the app descriptor, if required.
     if args.flash_args.image.check_app_descriptor.unwrap_or(true) {
-        check_bootloader(&elf_data)?;
+        check_idf_bootloader(&elf_data)?;
     }
 
     let mut monitor_args = args.flash_args.monitor_args;
